build(ssg): adjust ssg prerendered content embedding
Previously, prerendered pages included a JavaScript snippet to hide a splash
screen and global CSS styles directly. This change refactors the output by
moving the prerendered content styling into a `<noscript>` tag.

This ensures that the essential styling for the prerendered content is only
applied when JavaScript is disabled, providing a more robust base experience.
The JavaScript snippet for hiding the splash screen is removed from prerendered
pages as it is no longer necessary in this context.